baseinfosearch_1/requirements.txt
aiohttp
requests
beautifulsoup4
aiofiles
spacy
python-Levenshtein